Algebraic `RowEvaluator` (#3263)
In OVM witgen, we currenlty have (Symbolic)RowEvaluator, which is used
to remove range check side effects (as they are SymbolicExpression
based). We also use it to replay APC bus interactions, which are first
converted from AlgebraicExpression to SymbolicExpression. However, this
step is unnecessary, as this PR creates an (Algebraic)RowEvaluator which
we can use to evaluate APC bus interaction replay. This new row
evaluator is in the APC crate and can be used across different clients.