Split algebraic constraint modulo constant (#2751)
Splits constraints into multiple independent ones based on the following
idea:

Suppose we have the constraint `x + k * y + c = 0` with `x` and `y`
being variables (or expressions containing variables) and `k` and `c`
are constants. Furthermore, the range constraints of `x` and `y` are
such that no wrapping occurs in the operations, i.e. the constraint is
equivalent to the same constraint on the integers (where field elements
larger than half the modulus are mapped to negative numbers).

Then the same constraint is also true modulo `k`, where we get `x % k +
c % k = 0`. If this equation has a unique solution `s` in the range
constraints for `x`, we get a new constraint `x - s = 0`. We can
subtract that constraint from the original to get `k * y + c - s = 0`
and iterate.

Closes https://github.com/powdr-labs/powdr/issues/3108

---------

Co-authored-by: Thibaut Schaeffer <schaeffer.thibaut@gmail.com>