Remove bit decomposition. (#3243)
Removes the bit decomposition solver inside GroupedExpression /
AlgebraicConstraint.

Its functionality is now mostly taken over by the constraint splitter.
This is not fully true: The constraint splitter only works for offsets
that are constant, but this should not be a hindrance for our current
use-cases. If we need it, it should not be too difficult to extend it,
we only need to extend the final part of the `find_solution` function,
where it checks if the solution is unique.

---------

Co-authored-by: Thibaut Schaeffer <schaeffer.thibaut@gmail.com>