Remove the `components()` function (#3301)
In a review of https://github.com/powdr-labs/powdr/pull/3295 it became
obvious that the return type of `components()` is too complicated or at
least it is not really clear what it returns.

Instead of creating a struct, I split the function into three because
most of the time, we don't really need all three component types. In
addition to that, I changed the similar `into_components()` function to
return an iterator over an enum type that models the three kinds of
summands we have: quadratic, linear or constant because when we use that
function, we usually don't care about the individual "parts" but only
about their type.