Use unknown variables (#3304)
The first commit changes "referenced_variables" to
"referenced_unknown_variables" almost everywhere because this is also
what was meant. The difference between these two is that "known"
variables can occur inside expressions when they reference variables
that are known at that point in time during runtime solving. Nowadays,
there is no difference any more between these two concepts because we
only have FieldElement as the only implementation of RuntimeConstant.
But even with SymbolicExpression this is true.

The other uses of `referenced_variables` were in unused code, so I was
able to delete all of the code that used `referenced_variables` and with
it the trait that had that function.

Finally, `referenced_unknownw_variables` should be faster than
`referenced_variables` because it does not need to call the relevant
function on the runtime constants (which should always return an empty
iterator, but you never know what the optimizer can and can not do).