Make trace generation more GPU ready (#3356)
Currently, our `generate_trace` method which is supposed to be generic
uses an internal concrete type `Trace` to represent the traces of the
original instructions. This type is based on p3 which represents traces
as a contiguous piece of host memory.

To support GPU tracegen, this is not generic enough: once we call trace
gen for the original instructions, we receive some pointers to device
(GPU) memory. The plan is to then do something similar as on CPU: split
these original traces, grouping rows which correspond to the same apc
call, and map them to a single apc row on GPU.

This PR makes `generate_trace` generic over the type of the original
traces it is passed, with a flexible trait bound `TraceTrait` which only
exposes the width of the trace. For now, we also require `Send + Sync`,
because this works for the CPU traces. This might break with GPU, but
we'll see when we get there.