Test for APC not executed (#3359)
Currently our `guest_prove_simple` selects an APC that's not really
executed. This has been quite confusing when doing OVM 1.4 and GPU as we
were expecting APC to run while it doesn't. However, because we have
special trace gen path for when APC executed count is zero, this test is
still needed, but just under a different and more proper name, which
this PR does.

In addition, this PR also adds a new test that uses the instruction mode
instead to make sure APC is executed. This is the desired test that we
wanted to run in OVM 1.4 and GPU.