Introduce column allocator (#3387)
Alternative to #3385 which solves it in apc generation already.
The original bug is that when introducing a derived column (right now
only `is_valid`), the poly_id is chosen to be the max of the current
poly_ids in the machine plus one. But this poly_id may already be
pointed to by an original column in the substitutions, so in witgen it
looks like an apc column is both computed from an original column, and a
derived column.

This PR introduces a column allocator which issues new poly_ids that are
guaranteed not to be already used.
When returning the substitutions, we only keep the substitutions which
actually point to an apc column.