Remove max alignment from GPU (#3417)
Previously, in GPU, we calculate the estimate of APC calls to create
dummy arena in tracegen after subtracting `MAX_ALIGNMENT` from the
buffer length, because this is a "padding" OVM applies to buffers in
order to align memory address.

While this provides an exact # of APC calls, which is good, it creates a
maintenance risk in the future because this methodology can change in
the future, silently breaking our GPU prover.

We decided to remove this `MAX_ALIGNMENT` adjustment, because it's fine
to overestimate the dummy trace arena size, as the final trace gen use
the exact APC call count regardless.