[CI] GPU quick tests (#3433)
Added quick tests to GPU using our server as runner. However, instead of
building and running in the same job, uses a similar pipeline as CPU by
pretending that `build_gpu` and `test_quick_gpu` are run across
different runners, so that this can be adjusted easily once we have
proper runners.