Add pre_degree = 0 case when adding guard (#3460)
is_valid is a guard column to make all constraints and bus interactions
satisfied if needed. Adding it should not increase the degree of the
machine, but when the machine has degree 0 (no witness) but has bus
interactions, is_valid is added to the multiplicities of the bus
interactions, therefore make the degree of the machine increased from 0
to 1.