Put bigint branching instrutions in their own BB to not end up in APCs (#3464)
This is the simplest way of doing this:

- I kept the bigint branching instructions as branching instructions for
the sake of basic block detection, so that the instructions after them
go to a new BB as target.
- I removed bigint branching instructions from the allow list, which
ends up affecting BB detection, but I felt it was safer to keep both
flags for these instructions.

I have not tested this with the optimistic precompiles to see if it
fixes the bug.