Upgrade to OVM 1.4.2 (#3478)
Reth currently has AOT disabled, because PGO pc logging doesn't work for
AOT.

Should have fixed PC logging in TCO and non-TCO paths.

Tags should be final (which is used for all dependencies) but branch
name can be `v1.4.3`, so I will update them and re-tag next (which is
not related to this PR but clearer names for future updates).

---------

Co-authored-by: schaeff <thibaut@powdrlabs.com>