GPU mock prover (#3503)
With empirical constraint changes and `do_with_trace`, the GPU mock
prover path is removed, which I assume is because we don't want to deal
with the matrix generics (as I get `DeviceMatrix` error when I call
`do_with_trace` in `empirical_constraints.rs`)?

However, this mock prover path is still needed, especially because some
of the mock prover `powdr-openvm` tests are run in the GPU path as well.
Besides, the CLI `prove` also uses `do_with_trace` in the mock path,
which means that it will be incorrect if we run the CLI with `--features
cuda`.

This PR solves the issue by exposing another `do_with_cpu_trace`
function which shares helper functions with `do_with_trace`. The former
will always take the CPU mock path regardless of feature whereas the
latter takes a path as toggled by the feature flag.

This is achieved by making the `Sdk` generic over its config parameters
in the `do_with_trace_with_sdk` helper.