Rules for single-occurrence variables. (#3490)
Adds rules to replace certain variables by a single one.

The use-case here is for "diff_inv_marker_..." variables that each are
the inverse of certain variables only if those variables are non-zero
(and arbitrary otherwise). If the "diff_inv_marker_..." variables only
occur once, they are essentially "free" variables and under some
conditions, we can combine them into a single free variable and thus
reduce the number of variables.

Closes https://github.com/powdr-labs/powdr/issues/3370

---------

Co-authored-by: ShuangWu121 <47602565+ShuangWu121@users.noreply.github.com>