Substitute bus interaction fields if possible. (#3494)
Adds an optimizer step that goes through all bus interaction fields and
checks if they are constant. This situation is rather rare, but the
solver internally replaces all bus interaction fields by a new variables
(in the linearizer). Assignments to those new variables will not be
visible to the outside because the variables do not exist there. Most of
the time, though, assignments to them will lead to assignments to
actually existing variables. When this is not the case, this new step
asks the solver about the bus interaction field expressions and the
solver translates them to the new variables and thus can answer the
question via those.

---------

Co-authored-by: chriseth <chriseth.github@gmail.com>
Co-authored-by: Georg Wiese <georgwiese@gmail.com>